+++ src/components/AnimatePresence/PresenceChild.tsx
@@ -10,6 +10,7 @@ interface PresenceChildProps {
     onExitComplete?: () => void
     initial?: false | VariantLabels
     custom?: any
+    presenceAffectsLayout: boolean
 }
 
 let presenceId = 0
@@ -25,32 +26,43 @@ export const PresenceChild = ({
     isPresent,
     onExitComplete,
     custom,
+    presenceAffectsLayout,
 }: PresenceChildProps) => {
     const presenceChildren = useConstant(newChildrenMap)
     const id = useConstant(getPresenceId)
 
-    const context = useMemo(() => {
+    const context = useMemo(
+        () => {
+            return {
+                id,
+                initial,
+                isPresent,
+                custom,
+                onExitComplete: (childId: number) => {
+                    presenceChildren.set(childId, true)
+                    let allComplete = true
+                    presenceChildren.forEach((isComplete) => {
+                        if (!isComplete) allComplete = false
+                    })
+
+                    allComplete && onExitComplete?.()
+                },
+                register: (childId: number) => {
+                    presenceChildren.set(childId, false)
+                    return () => presenceChildren.delete(childId)
+                },
+            }
+        },
+        /**
+         * If the presence of a child affects the layout of the components around it,
+         * we want to make a new context value to ensure they get re-rendered
+         * so they can detect that layout change.
+         */
+        presenceAffectsLayout ? undefined : [isPresent]
+    )
+
+    useMemo(() => {
         presenceChildren.forEach((_, key) => presenceChildren.set(key, false))
-
-        return {
-            id,
-            initial,
-            isPresent,
-            custom,
-            onExitComplete: (childId: number) => {
-                presenceChildren.set(childId, true)
-                let allComplete = true
-                presenceChildren.forEach((isComplete) => {
-                    if (!isComplete) allComplete = false
-                })
-
-                allComplete && onExitComplete?.()
-            },
-            register: (childId: number) => {
-                presenceChildren.set(childId, false)
-                return () => presenceChildren.delete(childId)
-            },
-        }
     }, [isPresent])
 
     return (