+++ src/gestures/drag/VisualElementDragControls.ts
@@ -289,7 +289,10 @@ export class VisualElementDragControls {
         this.visualElement.resetTransform()
         this.visualElement.measureLayout()
         element.style.transform = transform
-        this.visualElement.rebaseTargetBox(true)
+        this.visualElement.rebaseTargetBox(
+            true,
+            this.visualElement.getBoundingBoxWithoutTransforms()
+        )
     }
 
     resolveDragConstraints() {